refactor(runtime): consolidate duplicate error classes with @seans-mfe/contracts
The six error classes under packages/runtime/src/errors/ duplicated
packages/contracts/src/errors/ byte-for-byte apart from comments
(constructor signatures and properties verified identical by diff), and
errors/index.ts was already a shim re-exporting from contracts. Delete
the redundant class files and redirect the two remaining deep imports
(base-mfe.ts ValidationError, timeout-wrapper.ts TimeoutError) to
@seans-mfe/contracts. Test imports of '../errors/TimeoutError' move to
the '../errors' barrel (path change only).

One source of truth for error classes; instanceof checks now always see
the contracts constructors.

Refs #253, #254, #260
